fixing a bug introduced in previous commit. We have to use doubles not floats when we convert GPU models to CPU, because CPU trained models use doubles. This introduced an error where a CPU checkpoint could not be loaded and used on CPU because the initial state was of float type. Anyone who has converted a model with the old script must rerun the conversion script so that double() CPU models are saved instead of float().

+++ convert_gpu_cpu_checkpoint.lua
@@ -64,7 +64,7 @@ protos = checkpoint.protos
 -- convert the networks to be CPU models
 for k,v in pairs(protos) do
 	print('converting ' .. k .. ' to CPU')
-	protos[k]:float()
+	protos[k]:double()
 end
 
 local savefile = opt.model .. '_cpu.t7' -- append "cpu.t7" to filename
